+++ src/actions/get-dependencies.js
@@ -1,7 +1,7 @@
 // @flow
 import { getDependencyTree } from '../dependency-builder';
 
-export default async function getDependenciesAction(baseDir, file, bindingPrefix): Promise<any> {
+export default async function getDependenciesAction(baseDir, file, bindingPrefix, resolveConfig): Promise<any> {
   const consumerPath = process.cwd();
-  return getDependencyTree(baseDir, consumerPath, [file], bindingPrefix);
+  return getDependencyTree({ baseDir, consumerPath, filePaths: [file], bindingPrefix, resolveModulesConfig: resolveConfig });
 }