+++ src/dependency-builder/dependency-tree/index.js
@@ -1,3 +1,4 @@
+import { isRelativeImport } from '../../utils';
 /**
 * this file had been forked from https://github.com/dependents/node-dependency-tree
 */
@@ -104,27 +105,20 @@ module.exports._getDependencies = function(config) {
   for (var i = 0, l = dependencies.length; i < l; i++) {
     var dep = dependencies[i];
     const isVue = (!Array.isArray(dep) && dep.isScript);
-    var result;
+    const cabinetParams = {
+      partial: dep,
+      filename: config.filename,
+      directory: config.directory,
+      ast: precinct.ast,
+      config: config.requireConfig,
+      webpackConfig: config.webpackConfig,
+      resolveConfig: config.resolveConfig,
+    };
     if (!Array.isArray(dep) && dep.isScript !== undefined) { // used for vue - return array of objects
-      result = cabinet({
-        partial: dep.dep,
-        filename: config.filename,
-        directory: config.directory,
-        ast: precinct.ast,
-        config: config.requireConfig,
-        webpackConfig: config.webpackConfig,
-        isScript: dep.isScript
-      });
-    } else {
-      result = cabinet({
-        partial: dep,
-        filename: config.filename,
-        directory: config.directory,
-        ast: precinct.ast,
-        config: config.requireConfig,
-        webpackConfig: config.webpackConfig
-      });
+      cabinetParams.partial = dep.dep;
+      cabinetParams.isScript = dep.isScript;
     }
+    const result = cabinet(cabinetParams);
     const dependency = isVue || dep.dep ? dep.dep : dep;
     if (!result) {
       debug('skipping an empty filepath resolution for partial: ' + dep);
@@ -147,13 +141,18 @@ module.exports._getDependencies = function(config) {
       debug('skipping non-empty but non-existent resolution: ' + result + ' for partial: ' + dep);
       continue;
     }
-    var pathMap = {dep: dependency , resolvedDep: result};
+    const pathMap = { importSource: dependency, resolvedDep: result };
     var importSpecifiersSupportedLang = ['es6', 'ts', 'stylable'];
     importSpecifiersSupportedLang.forEach((lang) => {
-      if (precinctOptions[lang] && precinctOptions[lang].importSpecifiers && precinctOptions[lang].importSpecifiers[dep]) {
-        pathMap.importSpecifiers = precinctOptions[lang].importSpecifiers[dep];
+      if (precinctOptions[lang] && precinctOptions[lang].importSpecifiers && precinctOptions[lang].importSpecifiers[dependency]) {
+        pathMap.importSpecifiers = precinctOptions[lang].importSpecifiers[dependency];
       }
     });
+    if (!isRelativeImport(dependency) && config.resolveConfig) {
+      // is includes also packages, which actually don't use customResolve, however, they will be
+      // filtered out later.
+      pathMap.isCustomResolveUsed = true;
+    }
 
     pathMapDependencies.push(pathMap);