+++ src/utils/index.js
@@ -3,6 +3,7 @@ import removeDirP from './remove-dir-p';
 import readFileP from './read-file-p';
 import isValidIdChunk from './is-valid-id-chunk';
 import isValidScopeName from './is-valid-scope-name';
+import isRelativeImport from './is-relative-import';
 
 export {
   writeFileP,
@@ -10,4 +11,5 @@ export {
   readFileP,
   isValidIdChunk,
   isValidScopeName,
+  isRelativeImport,
 };