+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [unreleased]
 
+## [0.10.17-dev.4] - 2018-05-17
+
 - support custom module resolution
 
 ## [0.10.17-dev.2] - 2018-05-15